+++ cypress/fixtures/dates.json
@@ -1,0 +1,3 @@
+{
+    "dates" : []
+}
