--- cypress/fixtures/dates.xml
@@ -1,41 +1,0 @@
-<response>
-    <header>
-        <resultCode>00</resultCode>
-        <resultMsg>NORMAL SERVICE.</resultMsg>
-    </header>
-    <body>
-        <items>
-            <item>
-                <dateKind>01</dateKind>
-                <dateName>추석</dateName>
-                <isHoliday>Y</isHoliday>
-                <locdate>20150926</locdate>
-                <seq>1</seq>
-            </item>
-            <item>
-                <dateKind>01</dateKind>
-                <dateName>추석</dateName>
-                <isHoliday>Y</isHoliday>
-                <locdate>20150927</locdate>
-                <seq>1</seq>
-            </item>
-            <item>
-                <dateKind>01</dateKind>
-                <dateName>추석</dateName>
-                <isHoliday>Y</isHoliday>
-                <locdate>20150928</locdate>
-                <seq>1</seq>
-            </item>
-            <item>
-                <dateKind>01</dateKind>
-                <dateName>대체공휴일</dateName>
-                <isHoliday>Y</isHoliday>
-                <locdate>20150929</locdate>
-                <seq>1</seq>
-            </item>
-        </items>
-        <numOfRows>10</numOfRows>
-        <pageNo>1</pageNo>
-        <totalCount>4</totalCount>
-    </body>
-</response>
