+++ cypress/integration/one_day_e2e.spec.js
@@ -1,8 +1,10 @@
 /// <reference types="cypress" />
 
+const holidays = ["20210801"];
+
 describe('end to end tests for a day', () => {
     beforeEach(() => {
-        cy.runServer();
+        cy.runServer(holidays);
         cy.visit('/');
     });
 
@@ -28,7 +30,7 @@ describe('end to end tests for a day', () => {
     });
 
     it('count 0 weekday for a holiday in the server', () => {
-        const date = new Date(2015, 8, 28);
+        const date = new Date(2021, 7, 1);
 
         cy.setDate(date);
         cy.hasResult(0);
