+++ package.json
@@ -10,8 +10,7 @@
     "react-date-range": "^1.1.3",
     "react-dom": "^17.0.2",
     "react-scripts": "^4.0.3",
-    "web-vitals": "^1.1.2",
-    "xml-js": "^1.6.11"
+    "web-vitals": "^1.1.2"
   },
   "scripts": {
     "start": "react-scripts start",
